requirements.txt
# GDELT Data Access
gdeltdoc==1.12.0
gdelt==0.1.14

# Data Processing
pandas>=2.0.0
pyarrow>=14.0.0
dask[complete]>=2023.0.0

# Environment & Scheduling
python-dotenv>=1.0.0
schedule>=1.2.0

# HTTP & Networking
requests>=2.31.0

# Database
# Using built-in sqlite3, no additional package needed

# Testing (optional, for development)
pytest>=7.4.0
pytest-cov>=4.1.0